refactor: Tab type type.ts로 분리
