+++ HDesign/src/components/Tab/Tab.type.ts
@@ -1,0 +1,8 @@
+export interface TabProps {
+  label: string;
+  content: React.ReactNode;
+}
+
+export interface TabsProps {
+  tabs: TabProps[];
+}
